refactor: 拆分 ToolConfigManager 以降低体积
